Update notifications table: split actions into individual buttons
