+++ src/components/admin/notifications/notifications-tabs.tsx
@@ -1,8 +1,8 @@
 import {Tabs, TabsContent, TabsList, TabsTrigger} from "@/components/ui/tabs";
 import {Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle} from "@/components/ui/card";
 import NotificationsTablePagination from "@/components/admin/notifications/notifications-table-pagination";
-import {Table, TableBody, TableCell, TableHead, TableHeader, TableRow} from "@/components/ui/table";
-import {Button} from "@/components/ui/button";
+import RecentNotificationsTable from "@/components/admin/notifications/recent-notifications-table";
+import ScheduledNotificationsTable from "@/components/admin/notifications/scheduled-notifications-table";
 
 const allNotifications = [
     {
@@ -127,108 +127,10 @@ const allNotifications = [
     }
 ]
 
-const getStatusColor = (status: string) => {
-    switch (status) {
-        case 'Delivered':
-            return 'text-green-600 border-green-400 bg-green-100'
-        case 'Scheduled':
-            return 'text-yellow-600 border-yellow-400 bg-yellow-100'
-        case 'Sending':
-            return 'text-orange-600 border-orange-400 bg-orange-100'
-        default:
-            return 'text-yellow-600 border-yellow-400 bg-yellow-100'
-    }
-}
+const recentNotifications = allNotifications.filter(notification => notification.status !== 'Scheduled');
+const scheduledNotifications = allNotifications.filter(notification => notification.status !== 'Delivered' && notification.status !== 'Sending');
 
-const renderRecentNotificationsTable = () => {
-    const recentNotifications = allNotifications.filter(notification => notification.status !== 'Scheduled');
-    if (recentNotifications.length === 0) {
-        return <p>No notifications match the current filters.</p>;
-    }
-    return (
-        <Table>
-            <TableHeader>
-                <TableRow>
-                    <TableHead className="font-semibold">Subject</TableHead>
-                    <TableHead className="font-semibold">Recipients</TableHead>
-                    <TableHead className="font-semibold">Type</TableHead>
-                    <TableHead className="font-semibold">Method</TableHead>
-                    <TableHead className="font-semibold">Status</TableHead>
-                    <TableHead className="font-semibold text-wrap whitespace-normal">Sent date & time</TableHead>
-                    <TableHead className="font-semibold">Actions</TableHead>
-                </TableRow>
-            </TableHeader>
-            <TableBody>
-                {recentNotifications.map((notification) => (
-                    <TableRow key={notification.notificationId}>
-                        <TableCell className="text-wrap whitespace-normal">{notification.subject}</TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.recipients}</TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.type}</TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.method}</TableCell>
-                        <TableCell>
-                            <span
-                                className={`px-2 py-0.5 border rounded-full font-semibold ${getStatusColor(notification.status)}`}>
-                                {notification.status}
-                            </span>
-                        </TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.dateTime}</TableCell>
-                        <TableCell>
-                            <Button variant={"outline"} size={"sm"}>{notification.actions}</Button>
-                        </TableCell>
-                    </TableRow>
-                ))}
-            </TableBody>
-        </Table>
-    )
-}
-
-const renderScheduledNotificationsTable = () => {
-    const scheduledNotifications = allNotifications.filter(notification => notification.status !== 'Delivered' && notification.status !== 'Sending');
-    if (scheduledNotifications.length === 0) {
-        return <p>No notifications match the current filters.</p>;
-    }
-    return (
-        <Table>
-            <TableHeader>
-                <TableRow>
-                    <TableHead className="font-semibold">Subject</TableHead>
-                    <TableHead className="font-semibold">Recipients</TableHead>
-                    <TableHead className="font-semibold">Type</TableHead>
-                    <TableHead className="font-semibold">Method</TableHead>
-                    <TableHead className="font-semibold">Status</TableHead>
-                    <TableHead className="font-semibold text-wrap whitespace-normal">Sent date & time</TableHead>
-                    <TableHead className="font-semibold">Actions</TableHead>
-                </TableRow>
-            </TableHeader>
-            <TableBody>
-                {scheduledNotifications.map((notification) => (
-                    <TableRow key={notification.notificationId}>
-                        <TableCell className="text-wrap whitespace-normal">{notification.subject}</TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.recipients}</TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.type}</TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.method}</TableCell>
-                        <TableCell>
-                            <span
-                                className={`px-2 py-0.5 border rounded-full font-semibold ${getStatusColor(notification.status)}`}>
-                                {notification.status}
-                            </span>
-                        </TableCell>
-                        <TableCell className="text-wrap whitespace-normal">{notification.dateTime}</TableCell>
-                        <TableCell>
-                            {notification.actions.split(',').map(action => (
-                                <Button key={action.trim()} variant={"outline"} size={"sm"} className="mr-2 last:mr-0">
-                                    {action.trim()}
-                                </Button>
-                            ))}
-                        </TableCell>
-                    </TableRow>
-                ))}
-            </TableBody>
-        </Table>
-    )
-}
-
-export default function NotificationsTable() {
+export default function NotificationsTabs() {
     return (
         <Tabs defaultValue="recent-notifications" className="w-full">
             <TabsList className="grid w-full grid-cols-4">
@@ -246,7 +148,9 @@ export default function NotificationsTable() {
                         </CardDescription>
                     </CardHeader>
                     <CardContent>
-                        {renderRecentNotificationsTable()}
+                        <RecentNotificationsTable
+                            recentNotifications={recentNotifications}
+                        />
                     </CardContent>
                     <CardFooter>
                         <NotificationsTablePagination/>
@@ -263,7 +167,9 @@ export default function NotificationsTable() {
                         </CardDescription>
                     </CardHeader>
                     <CardContent>
-                        {renderScheduledNotificationsTable()}
+                        <ScheduledNotificationsTable
+                            scheduledNotifications={scheduledNotifications}
+                        />
                     </CardContent>
                     <CardFooter>
                         <NotificationsTablePagination/>
